+++ libs/base_alloc/src/lib.rs
@@ -19,8 +19,8 @@ pub mod text;
 
 #[doc(hidden)]
 #[allow(unused_imports)]
-pub use all::*;
-pub mod all {
+pub use zall::*;
+pub mod zall {
     //! All the crate's items flat re-exported.
     //! <br/><hr>
     //!
@@ -45,9 +45,9 @@ pub use _workspace_internals::*;
 #[doc(hidden)] #[rustfmt::skip]
 pub mod _workspace_internals {
     #![allow(unused_imports)]
-    pub use devela_base_core::all::*;
+    pub use devela_base_core::zall::*;
     pub use devela_base_core::_workspace_internals::*;
 
-    pub use devela_base_num::all::*;
+    pub use devela_base_num::zall::*;
     pub use devela_base_num::_workspace_internals::*;
 }