+++ libs/base_core/src/media/mod.rs
@@ -1,6 +1,8 @@
 // devela_base_core::media
 //
 #![doc = crate::_DOC_MEDIA!()]
+#![doc = crate::_doc!(modules: crate; media: audio, color, draw, font, image, video)]
+#![doc = crate::_doc!(newline)]
 //
 // safety
 #![cfg_attr(base_safe_media, forbid(unsafe_code))]