+++ libs/base_core/src/num/mod.rs
@@ -1,6 +1,10 @@
 // devela_base_core::num
 //
 #![doc = crate::_DOC_NUM!()]
+#![doc = crate::_doc!(modules: crate; num: geom, logic, niche, ord, quant, rand)]
+#![doc = crate::_doc!(newline)]
+//!
+#![doc = crate::_doc!(extends: cmp, num)]
 //
 // safety
 #![cfg_attr(base_safe_num, forbid(unsafe_code))]