+++ libs/base_core/src/phys/mod.rs
@@ -1,6 +1,10 @@
 // devela_base_core::phys
 //
 #![doc = crate::_DOC_PHYS!()]
+#![doc = crate::_doc!(modules: crate; phys: bio, chem, elec, mech, time, unit, wave)]
+#![doc = crate::_doc!(newline)]
+//!
+#![doc = crate::_doc!(extends: time)]
 //
 // safety
 #![cfg_attr(base_safe_phys, forbid(unsafe_code))]