+++ libs/base_num/src/lib.rs
@@ -16,8 +16,8 @@ mod float; // Float, FloatConst
 mod quant; // ValueQuant
 
 #[doc(hidden)]
-pub use all::*;
-pub mod all {
+pub use zall::*;
+pub mod zall {
     //! All the crate's items flat re-exported.
     //! <br/><hr>
     //!
@@ -42,7 +42,7 @@ mod _crate_internals {
     #[rustfmt::skip]
     pub use {
         devela_base_core::_workspace_internals::*,
-        devela_base_core::all::*,
+        devela_base_core::zall::*,
     };
 }